elm-core: update subtle references to maybe.maybe
The thing: erl2ocaml assumes that the current module is correct wrt to local definition. So, eg, in list module `maybe.nothing{}` was considered as a local enum_ctr, however, - because maybe.mli was "opened" (because of usages of types from it), - the constructors were imported from it, - so ocamlc was OK for compilation because of this subtlety.