Create annotations correctly for binary syntax
This handles that some syntax parts e.g. lists of modifiers can be empty and thus have no annotations, and also that some constructs in the binary syntax do not have any annotations even though they have a position in the file, we should probably change this in the future, for now we strip the annotations out after using them to get the correct position for the surrounding expression.

This commit also adds a test case, with binary expressions of various kinds, including both construction and matching