Several annotation fixes.
-For binary operations: the position is now the whole operation not the position of the operator itself.

-For unary operations: the position is now the whole operation not the position of the operator itself.

-For map expression: The position of the map tuple is now part of the position of the whole map expression

-For map types: Both => and := operations are treated the same with respect to position, the position is  the position ofg the whole association

-Pattern argument lists: The parenthesis are the starts and ends of the position of the pattern argument list.

-For an expression in parentheses: use the location of the parantheses as start and end.

-for attributes include the whol;e attribute including the - in the position of the attribute

Updating expected error messages.

Since we changed the positions for expressions the error messages have changed. The main cause for this is including the - as the start of an attribute. This is debatable. The other changes is because of a binary operation now including the operands in one case this is a big improvement, in the other were the operator is explicit ly mentioned in the error message it is less clear.