Introduce erltc_ast
It offers an abstraction around traversing the AST similar to
Elixir's Macro.prewalk/postwalk/traverse.

This is also similar to OTP's erl_syntax, but does not try to abstract
around AST structure allowing to directly pattern match on the values
instead of using function-based constructors and desctructors, that
should be much more lightweight in practice (though less robust).