Add linting for globally defined types, in the scan phase, when we do not have access to these global definitions we assume usages are correct, and only verify them in the compile phase.
Add test cases for lint failures, also make sure error messages for enums to print a.b rather than {a,b} for remote enums. The change required adding a dep directory to checks in order to hve a definitions that can be used in the examples to provoke certain cross module lint issues.