erlbuild: replace parse_transform with pre compile hook & escript
Replace parse_transform for reading Makefile template with a compile
hook and escript that generates .erl file with the contents of the
template.

This reduces complexity, removes largely unnecessary parse transform,
and allows to build erlbuild from a different directory path.

Immediate motivation for this change is enabling erlbuild integration
with rebar3 plugin which requires building it from a different path.