Add unchecked and opaque modifiers to the language.
I've taken a new approach to this where the modifiers are part of the same form as the things they modify

This allows writing:

[opaque, unchecked]
-type x() :: term()

To specify a type which is opaque in static contexts with a dynamic definition. It also makes it possible to write:
[unchecked]
function(A,B) -> [A,B].

Meaning that the function should not be type checked.