Convert workflows to Dahl (pre-work)
This change should make no difference to behavior, but is prep
for follow-on work.

We generate the GitHub workflow yamls using dhall.

This has the following nice properties:
- less repetition in sterlang.yml, so less risk of things
falling out of sync
- in follow-ons, we can parameterize things if we want
- the types caught a minor bug. In erlt.yml, we specified
'strategy' without 'matrix', which has no effect.
https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-syntax-for-github-actions#jobsjob_idstrategy
to do a similar-but-different job on release as compared to push.

And the following downsides:
- editing the dhall is only nice with a bunch of plugins installed (see
README)
- the config got slightly *longer* for now, though I think the
light abstraction will soon pay its cost