First pass at ErlT syntax highlighting support for VSCode
It is based on the erlang_ls/vscode one, with very minor package name
tweaks to prevent a clash with the existing one.

Its only addition at this stage is to include ".erlt" as an extension,
mapped also to erlang.

Next step is to either separate a language id to erlt, or to make the
syntax grammar aware of both, and treat erlt as we decide.