Add .etf and .defs listings to erltc (#290)
This can be used for easy testing of the integration with sterlang.

They can be achieved with `+etf` or `+defs` flags. The output
file is produced in the compiler's output directory (CWD by default,
can be controlled with `-o` option) with the same name as the module 
and a `.etf` or `.defs` extension.