Separate nullary variants from variants with no fields
If a variant is defined without fields `foo`, it has to be used
without fields: `foo`, not `foo{}`. If it is defined with fields,
even if empty: `foo{}`, it has to be used with fields `foo{}`, not
`foo`.

This modifies AST of variants to include `none` as possible definition
or usage of fields.