Fix #329 Classic Erlang type emit for shapes
Fix #330
Fix #329

Before, for shape variables, we  generating
classic Erlang type specs with unused variables.

For specs, we can fix this with guards.

-spec foo(A) -> .... where A :: #{atom() -> any()}

For other types (which don't have guards),
if the type variable for the shape is unused
in the generated Erlang, then replace it with
__ERLT_SHAPE_VAR_<old name here>. The leading `_`
makes the generated code pass classic Erlang erl_lint.

If there is a conflict with an existing variable,
we prepend more underscores until there is no conflict.

> See examples/dev_struct/src/shape_mod2.erlt for
each of these cases.