Infra for built-in defs and runtime (hack for M0)
write defs files for built_in types and modules for the erlt runtime.

- built_in defs are written to the build dir. This PR includes defs for
a single function from the lists module, just to get us started.
- modules for the erlt runtime are written directly into the source dir.
This PR includes t_io:format/0 and t_io:format/1.

This is a hack for our M0 milestone.

The build model for the erlt runtime and built-in defs is TBD. We
will likely want the support to piggyback on support for
erlt builds with dependencies.

See erltc/built_ins/README.md for how to update built_ins