Make AST representation for catch clauses more explicit
Instead of always containing a 3-tuple with Kind, Reason, Stack,
switch to containing a head with 3 possible slots for each of those.

The primary goal is differentiating between `catch X` and `catch throw:X`.

Only allow catches without class or stacktrace capturing in checked code,
due to issues with typing the class and stacktrace. This is verified
in erltc, so sterlang never deals with unsupported cases.