erltc finds sterlang or errors
> some of the details below are in
follow-on commits in the same PR

user-facing behavior, **assuming sterlang is installed**:
- just works, assuming the user is on rebar 3.14.2 or
is using scripts/rebar3

(see scripts/README.md for what we did here)

Stuff for our own development:
- erltc errors if it can't find sterlang
- Running 'make' in examples still connects to the daemon, but
it also installs the sterlang jar so that subsequent runs of
`rebar3 compile` and `rebar3 shell` will do the right thing
(sterlang will be found)
- erltc --build requires sterlang
- when running erltc without `--build` (single-file mode),
`--skip-type-checking` can be passed. This is a hack to
get our checks tests to pass

Stuff for ci:
- When there is a push with a change to erltc, erltc tests are run with the
sterlang jar.
- When there is a push with a change to sterlang, erltc tests are run
with both the jar and the native sterlang.
- When there is a release, erltc tests are run with both the jar and the
native sterlang binary.