It is not a drop in replacement for gb_trees, because it returns an option enum, when that is appropriate, the internal representation of the tree is also different, and less compact.
The change adds the max function to the erlang module. It also adds an option type to the erlang module, and introduce a hack where a dependency on a type from the erlang module does not generate a dependency.

One thing that took a lot of time was that gb_trees was using an untagged tuple in an option type (in order to save space probably) and this made it really hard to find all the places you needed to change the four tuple into an enum constructor.

Other issues:

Changed erlang:error so it accepts a term(). All guard functions aren't typed automatically which is annoying, but in this case they could all be taken away, and maybe that is true in general i.e. all is_integer etc.