Minimal docs
Copy minimal specs for enums, structs, and shapes to doc/

Add to the doc/README.md so it covers all of the basics of ErlT
not covered by the enums, structs, and shapes docs.

Delete the language overview, since it was only a few sentences
and overlapped with doc/README

Add qualification to IDE support about workspace root.

Update the README template so it has the right title and links to IDE
support.