Hack IDE support so 'play' works out of the box
Do the dumb hardcoding needed so that
the examples in 'play' and any new erlt apps
made in 'play' will get a nice IDE experience
out of the box.

This is not the best solution long-term, but
it is the most pragmatic way to make sure to
get our (exciting!) IDE experience in front of users.

Steps to test:

create and install a vsix using the instructions in
<repo_root>/vscode.

Be sure you have a sterlang.jar or sterlang in
<repo_root>/erltc/priv. See README for how to do this.

Then: Doesn't matter what your vscode workspace is: open
any erlt file under <repo_root>/play. Make a type error
and ensure the IDE quickly yells at you.

Also, ensure that the IDE experience still works
for 'tests'. For 'tests', you need to have
the workspace root actually be 'tests'. This annoying
restrictioni is because we use the sterlang daemon
when developing tests and have plumbing for that in tests/Makefile,
so we need a custom command, which then needs to be fs aware and
blah blah blah blah blah