[elp] set language level to 11 in generated tree-sitter parser
And change the language name to erlang_elp, so we can keep compatible
with the classic erlang too, as likely to be published by
@AbstractMachinesLab.

The tree-sitter-cli dep is downgraded to 0.16.2 so the generated
parser is compatible with the current haskell tree-sitter support.

This can probably be updated via the appropriate PR on
haskell-tree-sitter, at t later date.