[h-elp] Incremental parsing (#496)
A PR to essentially capture my work in progress.

It builds a language server that can parse an .elp file based on the test tree-sitter grammar in the elp/tree-sitter-erlang directory.

This can produce a well-typed AST derived from the grammar using Template Haskell by semantic. The AST is not used for anything at present.

It also generates as-you-type diagnostics for the .elp file, based on parsing the tree-sitter grammar.