[elp] Adding draft passthru connection for erlang_lsp
This is just initial hacking, I *think* it establishes a context where
we can have a sub-process with wrapped LSP Message streams for its io,
and have the same from the LSP client (the IDE).

If this works, first step is to just forward messages throught the
pairs, logging for interest.

Very complicated tee comand.